Integrated Plotting and Natural Random Example
[Natural.java] Updated the natural test function to
take in a genome and then pick random positions on it
extract a subsequence of length 0-30 and pluck it
out of the genome.

From here the new genome is saved to disk and
an oligo pool of insertions created with the intent to
insert them right back into the genome we just made.

[Plot.java] This file is taken from the easyplot project
Plot is a light weight easy plotting wrapper for GNUPLOT

[test.FASTA.java]  This test is for testing and verifying
the FASTA tool.

[util.TextFile.java] Added a string splitter function to this
library

[Heuristic.java] Added some more print statements for context